Add float instead of int

+++ hw-2/engine.py
@@ -3,5 +3,5 @@ from dataclasses import dataclass
 
 @dataclass
 class Engine:
-    volume: int = 2.7
+    volume: float = 2.7
     piston: int = 4
